+++ mpihans.py
@@ -1,16 +1,15 @@
-from re import A
 from timeit import default_timer as timer
 import sys
 from mpi4py import MPI
 import NS_hsa as NS
-from hans_io import read_hans_file
+import hans_io
 from numpy.random import MT19937
 from numpy.random import RandomState, SeedSequence
 import os
 import ase
 import h5py
 import numpy as np
-import cProfile
+#import cProfile
 
 comm = MPI.COMM_WORLD
 size = comm.Get_size()
@@ -27,9 +26,8 @@ def main():
     dh_min = 1e-10
     dshr_min = 1e-5
     dstr_min = 1e-5
+    min_dstep = np.array([dv_min,dr_min,dt_min,dh_min,dshr_min,dstr_min])
 
-    lower_bound = 0.2 #mc adjust upper and lower bounds
-    upper_bound = 0.5
 
     mc_adjust_wl = max(10//size,1)
 
@@ -41,7 +39,7 @@ def main():
     args = {}
     from_restart = None
     if rank == 0:
-        args = read_hans_file() #parsing input file
+        args = hans_io.read_hans_file() #parsing input file
         if "from_restart" in args and int(args["from_restart"]):
                 print(f"Attempting to restart from {directory}")
                 from_restart = True
@@ -84,22 +82,6 @@ def main():
         args["min_angle"] = float(args["min_angle"])
         args["min_aspect_ratio"] = float(args["min_aspect_ratio"])
         
-
-
-    # if not from_restart:
-    #     if rank == 0:
-    #         if "directory" in args:
-    #             dir_prefix = args["directory"]
-    #         else:
-    #             dir_prefix = f"NeSa_{args['nchains']}_{args['nbeads']}mer.{args['nwalkers']}.{args['walklength']}"
-    #             i_n = 1
-    #         while os.path.exists(f"{dir_prefix}.{i_n}/"):
-    #             i_n += 1
-    #         directory = f"{dir_prefix}.{i_n}/"
-
-    # print(f"Creating directory {directory} to output results")
-    # os.mkdir(f"{directory}")
-    # os.chdir(f"{directory}")
 
     if rank == 0:
         for arg in args:
@@ -191,8 +173,8 @@ def main():
 
         if rank == vol_max_index[0]:
             if i%100 == 0:
-                NS.write_to_extxyz(args,vol_max_index[1]+1, filename=f"traj.extxyz")
-                print(i)
+                hans_io.write_to_extxyz(args,vol_max_index[1]+1, filename=f"traj.extxyz")
+                print(i, vol_max)
             active_walker = vol_max_index[1]
             NS.import_ase_to_ibox(config_to_clone,active_walker+1,args)
         else:
@@ -207,51 +189,9 @@ def main():
 
 
         if i%mc_adjust_interval == 0:
+            NS.adjust_mc_steps_2(args,comm,move_ratio,vol_max,walklength = mc_adjust_wl, 
+                      min_dstep=min_dstep, dv_max=dv_max,dr_max=dr_max,dshear = dshear, dstretch = dstretch)
             #print(NS.alk.alkane_get_dv_max(),NS.alk.alkane_get_dr_max(),dshear,dstretch, rank)
-            rate = np.zeros(6)
-            avg_rate = np.zeros_like(rate)
-            mc_box = np.random.randint(args["nwalkers"])
-
-            for i in range(6):
-                move_ratio_matrix = np.eye(6)
-                backup = NS.mk_ase_config(mc_box+1,args["nbeads"],args["nchains"],scaling=1)
-                if move_ratio[i] != 0:
-                    rate += NS.MC_run_2(args,mc_adjust_wl, move_ratio_matrix[i],mc_box+1,vol_max,dshear=dshear, dstretch=dstretch)[1]
-                    NS.import_ase_to_ibox(backup,mc_box+1,args)
-            comm.Allreduce(rate,avg_rate,op=MPI.SUM)
-            avg_rate = avg_rate/size
-            if move_ratio[0] != 0:
-                if avg_rate[0] < lower_bound:
-                    NS.alk.alkane_set_dv_max(max(0.5*NS.alk.alkane_get_dv_max(),dv_min))
-                elif avg_rate[0] > upper_bound:
-                    NS.alk.alkane_set_dv_max(min(2.0*NS.alk.alkane_get_dv_max(),dv_max))
-            if move_ratio[1] != 0:
-                if avg_rate[1] < lower_bound:
-                    NS.alk.alkane_set_dr_max(max(0.5*NS.alk.alkane_get_dr_max(),dr_min))
-                elif avg_rate[1] > upper_bound:
-                    NS.alk.alkane_set_dr_max(min(2.0*NS.alk.alkane_get_dr_max(),dr_max))
-            if move_ratio[2] != 0:
-                if avg_rate[2] < lower_bound:
-                    NS.alk.alkane_set_dt_max(max(0.5*NS.alk.alkane_get_dt_max(),dt_min))
-                elif avg_rate[2] > upper_bound:
-                    NS.alk.alkane_set_dt_max(2.0*NS.alk.alkane_get_dt_max())
-            if move_ratio[3] != 0:
-                if avg_rate[3] < lower_bound:
-                    NS.alk.alkane_set_dh_max(max(0.5*NS.alk.alkane_get_dh_max(),dh_min))
-                elif avg_rate[3] > upper_bound:
-                    NS.alk.alkane_set_dh_max(2.0*NS.alk.alkane_get_dh_max())
-            if move_ratio[4] != 0:
-                if avg_rate[4] < lower_bound:
-                    dshear = max(0.5*dshear,dshr_min)
-                elif avg_rate[4] > upper_bound:
-                    dshear = 2.0*dshear
-            if move_ratio[5] != 0:
-                if avg_rate[5] < lower_bound:
-                    dstretch = max(0.5*dstretch,dstr_min)
-                elif avg_rate[5] > upper_bound:
-                    dstretch  = 2.0*dstretch
-
-
     if rank == 0:
         f.close()
 
@@ -259,47 +199,13 @@ def main():
     if rank == 0:
         print("NS RUN TIME TAKEN =", t2-t0)            
         print("writing restart")
-    if rank ==0:
-        f = h5py.File("restart.hdf5", "w")
-        for j in args:
-            f.attrs.create(j,args[j])
-        f.attrs.create("prev_iters",i+1)
-        for iwalker in range(1,args["nwalkers"]+1):
-            groupname = f"walker_{rank}_{iwalker:04d}"
-            tempgrp = f.create_group(groupname)
-            coords = tempgrp.create_dataset("coordinates",(args["nbeads"]*args["nchains"],3),dtype="float64")
-            unitcell = tempgrp.create_dataset("unitcell",(3,3),dtype="float64")
-
-            unitcell[:] = NS.alk.box_get_cell(iwalker)
-            for ichain in range(args["nchains"]):
-                chain = NS.alk.alkane_get_chain(ichain+1,iwalker)
-                coords[ichain*args["nbeads"]:ichain*args["nbeads"]+(args["nbeads"]), :] = chain
-        for j in range(1,size):
-            print(j, " we start here")
-            config_list = comm.recv(source=j,tag = j)
-            for iwalker in range(1,args["nwalkers"]+1):
-                groupname = f"walker_{j}_{iwalker:04d}"
-                tempgrp = f.create_group(groupname)
-                coords = tempgrp.create_dataset("coordinates",(args["nbeads"]*args["nchains"],3),dtype="float64")
-                unitcell = tempgrp.create_dataset("unitcell",(3,3),dtype="float64")
-
-                unitcell[:] = config_list[iwalker-1].cell
-                for ichain in range(args["nchains"]):
-                    coords[:] = config_list[iwalker-1].positions
-        f.close()
-    else:
-        config_list=[NS.mk_ase_config(ibox+1,args["nbeads"],args["nchains"],1.0) for ibox in range(args["nwalkers"])]
-        comm.ssend(config_list,0,tag=rank)
+    hans_io.write_to_restart(args,comm,filename = "restart.hdf5",i=0)
 
     sys.stdout.flush()
     NS.alk.alkane_destroy()
     NS.alk.box_destroy()
     comm.Barrier()
 
-    # print(NS.alk.alkane_get_dv_max(), "dv")
-    # print(NS.alk.alkane_get_dr_max(), "dr")
-    # print(NS.alk.alkane_get_dt_max(), "dt")
-    # print(NS.alk.alkane_get_dh_max(), "dh")
 if __name__ == "__main__":
     # prof = cProfile.Profile()
     # prof.enable()
